+++ packages/nestjs-traceable/src/nestjs/utils/index.ts
@@ -1,0 +1,9 @@
+// Error 처리 유틸리티
+export {
+  getErrorMessage,
+  getErrorDetails,
+  getErrorChain,
+  getRootCause,
+  getFullErrorDetails,
+  formatErrorForLogging,
+} from './error.util';